feat: consolidate and refactor codebase for cleaner structure
- Delete the entire contents of `app copy.tsx`
- Add a new file `index.tsx` in the `Finished` directory with 1641 lines of code
- Add a new file `style.scss` in the `Finished` directory with 77 lines of code
- Add a new file `index.ts` in the `views` directory with 4 lines of code
