+++ src/views/Finished/style.scss
@@ -1,0 +1,77 @@
+.ded-map-container:has(:first-child),
+.ded-map-container:has(:last-child) {
+  padding-block: 0;
+
+  .ded-card-body {
+    padding-inline: 0;
+  }
+}
+
+.ded-divider-horizontal.ded-divider::before,
+.ded-divider-horizontal.ded-divider::after {
+  border-top-color: #EFEFEF;
+}
+
+.ded-source-container {
+  gap: 0;
+
+  &:has(:first-child),
+  &:has(:last-child) {
+    padding-block: 8px;
+  }
+}
+
+.flip-card {
+  position: fixed;
+  right: 16px;
+  bottom: 16px;
+  perspective: 1000px;
+}
+
+.flip-card-inner {
+  position: relative;
+  width: 80px;
+  height: 80px;
+  text-align: center;
+  transition: transform 0.6s;
+  transform-style: preserve-3d;
+  animation: initialFlip 1s ease-in-out;
+}
+
+.flip-card:hover .flip-card-inner {
+  transform: rotateY(180deg);
+}
+
+.flip-card-front,
+.flip-card-back {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  backface-visibility: hidden;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: white;
+  font-size: 1.5rem;
+  border-radius: 15px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.2);
+}
+
+.flip-card-front {
+  background: #007acc;
+}
+
+.flip-card-back {
+  background: #003f5c;
+  transform: rotateY(180deg);
+}
+
+@keyframes initialFlip {
+  from {
+    transform: rotateY(0deg);
+  }
+
+  to {
+    transform: rotateY(720deg);
+  }
+}
