+++ src/views/index.ts
@@ -1,0 +1,4 @@
+import { Demo } from './Demo';
+import { Finished } from './Finished';
+
+export { Demo, Finished };
